Plumb in functionality to choose keychain.
This functionality allows access to keychain of choice, the primary reason
is to target System keychain as a place to store items.
It consciously uses a deprecated SecKeychainOpen call, because that's,
at present, the only way to access that keychain.

It is deprecated, though the DTS at apple forums suggest something
similar to this for LaunchDaemons to have access to the keychain.